src/test/java/dev/turingcomplete/asmtestkit/assertion/representation/AttributeRepresentationTest.java
package dev.turingcomplete.asmtestkit.assertion.representation;

import dev.turingcomplete.asmtestkit.assertion.__helper.DummyAttribute;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class AttributeRepresentationTest {
  // -- Class Fields ------------------------------------------------------------------------------------------------ //
  // -- Instance Fields --------------------------------------------------------------------------------------------- //
  // -- Initialization ---------------------------------------------------------------------------------------------- //
  // -- Exposed Methods --------------------------------------------------------------------------------------------- //

  @Test
  void testNullArgument() {
    assertThat(AttributeRepresentation.instance().toStringOf(null))
            .isEqualTo(null);
  }

  @Test
  void testStringRepresentation() {
    assertThat(AttributeRepresentation.instance().toStringOf(new DummyAttribute("Name", "Content")))
            .isEqualTo("NameContent");
  }

  // -- Private Methods --------------------------------------------------------------------------------------------- //
  // -- Inner Type -------------------------------------------------------------------------------------------------- //
}
